refactor: js -> jsx, unused 정리, require -> import
