+++ screens/auth/password/ResetPassword.jsx
@@ -6,38 +6,36 @@ import {
   Image,
   TextInput,
   TouchableOpacity,
-  ScrollView
+  ActivityIndicator,
 } from "react-native";
 import { requestApi } from "../../../utils/apiSetting";
-import { useDispatch } from "react-redux";
-import { setAccessToken, setUserName } from "../../../store/reducers/user-slice";
+import logoImage from "../../../assets/icon.png";
+import emailIcon from "../../../assets/email.png";
+import passwordIcon from "../../../assets/password.png";
 
-const logoImage = require("../../../assets/icon.png");
-const emailIcon = require("../../../assets/email.png");
-const passwordIcon = require("../../../assets/password.png");
-
-const LoginScreen = ({ navigation }) => {
-  const [loginForm, setLoginForm] = useState({
+const ResetPassword = ({ navigation }) => {
+  const [resetPasswordForm, setResetPasswordForm] = useState({
+    username: "",
     email: "",
-    password: "",
   });
-
-  const dispatch = useDispatch();
+  const [isLoading, setIsLoading] = useState(false);
 
   const handleLogin = () => {
+    setIsLoading(true);
     requestApi
-      .post("/api/auth/login", {
-        email: loginForm.email,
-        password: loginForm.password,
+      .post("/api/auth/reset-password", {
+        username: resetPasswordForm.username,
+        email: resetPasswordForm.email,
       })
       .then((res) => {
-        dispatch(setUserName(res.data.result.username));
-        dispatch(setAccessToken(res.data.result.accessToken));
-        navigation.navigate("MainPage");
+        navigation.navigate("Login");
         console.log(res.data);
       })
       .catch((err) => {
         console.log(err);
+      })
+      .finally(() => {
+        setIsLoading(false);
       });
   };
 
@@ -45,15 +43,15 @@ const LoginScreen = ({ navigation }) => {
     <View style={styles.container}>
       <View style={styles.contentContainer}>
         <Image source={logoImage} style={styles.logo} />
-        <Text style={styles.title}>로그인</Text>
+        <Text style={styles.title}>비밀번호 초기화</Text>
 
         <View style={styles.inputContainer}>
           <TextInput
-            value={loginForm.email}
+            value={resetPasswordForm.username}
             onChangeText={(text) =>
-              setLoginForm((prev) => ({
+              setResetPasswordForm((prev) => ({
                 ...prev,
-                email: text,
+                username: text,
               }))
             }
             style={styles.input}
@@ -65,36 +63,36 @@ const LoginScreen = ({ navigation }) => {
 
         <View style={styles.inputContainer}>
           <TextInput
-            value={loginForm.password}
+            value={resetPasswordForm.email}
             onChangeText={(text) =>
-              setLoginForm((prev) => ({
+              setResetPasswordForm((prev) => ({
                 ...prev,
-                password: text,
+                email: text,
               }))
             }
             style={styles.input}
-            placeholder="Password"
+            placeholder="Email"
             placeholderTextColor="#787878"
-            secureTextEntry
           />
           <Image source={passwordIcon} style={styles.iconp} />
         </View>
 
-        {/*<TouchableOpacity style={styles.button} onPress={() => navigation.navigate('MainPage')}>*/}
-        <TouchableOpacity style={styles.button} onPress={handleLogin}>
-          <Text style={styles.buttonText}>로그인</Text>
+        <TouchableOpacity
+          style={styles.button}
+          onPress={handleLogin}
+          disabled={isLoading}
+        >
+          {isLoading ? (
+            <ActivityIndicator size="small" color="#FFFFFF" />
+          ) : (
+            <Text style={styles.buttonText}>초기화</Text>
+          )}
         </TouchableOpacity>
 
         <View style={styles.linksContainer}>
-          <TouchableOpacity
-            onPress={() => navigation.navigate("ResetPassword")}
-          >
-            <Text style={styles.link}>비밀번호 초기화</Text>
+          <TouchableOpacity onPress={() => navigation.navigate("Login")}>
+            <Text style={styles.link}>로그인 페이지로</Text>
           </TouchableOpacity>
-          {/*<Text style={styles.linkDivider}> | </Text>*/}
-          {/*<TouchableOpacity>*/}
-          {/*    <Text style={styles.link}>비밀번호 찾기</Text>*/}
-          {/*</TouchableOpacity>*/}
         </View>
       </View>
 
@@ -132,7 +130,7 @@ const styles = StyleSheet.create({
   title: {
     fontSize: 30,
     fontWeight: "bold",
-    color: "#22215B",
+    color: "#333333",
     marginBottom: 40,
   },
   inputContainer: {
@@ -157,9 +155,9 @@ const styles = StyleSheet.create({
     marginRight: 10,
   },
   iconp: {
-    width: 20,
+    width: 28,
     height: 28,
-    marginRight: 10,
+    marginRight: 5,
   },
   button: {
     width: "80%",
@@ -212,4 +210,4 @@ const styles = StyleSheet.create({
   },
 });
 
-export default LoginScreen;
+export default ResetPassword;
